Build changes and preliminary ncurses printing
Build is still kind of finnicky. Right now you need to make sure you
build libdemise-core in demise-core/build and then make install, which
will install libdemise-core.a in demise-core/lib, before you make
demise-ncurses because CMake will look for libdemise-core.a in
demise-core/lib. If demise-core is not built before demise-ncurses,
CMake will either not find libdemise-core.a, or if it doesn't exist,
fail to build.

We can now print the first 10 lines of a file.